Addressed review: MA-reachability, track-identity guard, mirror split
